pyproject.toml
[tool.ruff]
line-length = 88

# Enable Pyflakes `E` and `F` codes by default.
select = ["E", "F"]
ignore = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

per-file-ignores = {}


# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

target-version = "py310"

[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[tool.poetry]
name = "note-block-flask"
version = "0.1.0"
description = ""
authors = ["Pedro Fonseca <contatopedrorn@gmail.com>"]
readme = "README.md"
packages = [{include = "note_block_flask"}]


[tool.poetry.dependencies]
python = "^3.8"
flask = "^2.2.2"
sqlmodel = "^0.0.8"
psycopg2 = "^2.9.5"


[tool.poetry.group.dev.dependencies]
pre-commit = "^2.21.0"
ruff = "^0.0.207"
pytest = "^7.2.0"
python-dotenv = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
